+++ CHANGELOG.md
@@ -11,7 +11,9 @@ All notable changes to this project will be documented in this file.
 ### Feature
 
 - [x] **$vuepress-plugin-mathjax**: render TeX formula into PNGs at server side
-- [ ] **$markdown-it-diagrams**: support diagrams listed in [here](https://shd101wyy.github.io/markdown-preview-enhanced/#/diagrams)
+- [x] **$vuepress-plugin-diagrams**: render diagrams to PNGs at server side
+- [x] **$vuepress-plugin-diagrams**: support add classes and styles to graphs
+- [ ] **$vuepress-plugin-diagrams**: support diagrams listed in [here](https://shd101wyy.github.io/markdown-preview-enhanced/#/diagrams)
 - [ ] reimplement medium zoom
 - [ ] update sass loader (wait for VuePress to solve issue #2148)